refactor: Replace `AddVulnerabilitiesCommand` with unified `AddCommand`
- Merged functionality of `AddVulnerabilitiesCommand` into `AddCommand` for a streamlined CLI experience.
- Updated tests, docs, and examples to reflect the new `add` command structure.
- Simplified usage by removing redundancy in the command hierarchy.
