+++ docs/modules/ROOT/pages/cli-overview.adoc
@@ -63,7 +63,7 @@ Available on `report` and `suppress` commands.
 | xref:cli-report.adoc[report]
 | Generate an HTML vulnerability report.
 
-| xref:cli-add.adoc[modify add vulnerability]
+| xref:cli-add.adoc[modify add]
 | Add a new vulnerability entry to one or more Vulnlog files.
 
 | xref:cli-copy.adoc[modify copy]
