+++ website/index.html
@@ -302,7 +302,7 @@
                     Vulnlog file created at: vulnlog.yaml
 
                     # Add a vulnerability finding from the command line
-                    $ vulnlog modify add vulnerability vulnlog.yaml --vuln-id CVE-2026-1234 --reporter trivy
+                    $ vulnlog modify add vulnlog.yaml --vuln-id CVE-2026-1234 --reporter trivy
                     Added to vulnlog.yaml: CVE-2026-1234
 
                     # Validate your Vulnlog files
